s390x: sign-extend the BRCTH displacement from 32 bits
BRCTH is RIL-b, so its I2 field is a signed 32-bit count of halfwords, but
the handler sign-extended it from 16 bits. A displacement outside
+/-32Ki halfwords was truncated, and the lifter emitted a branch to the
wrong address with no crash and no warning. The 0x10000-halfword case
lands back on the instruction's own address, which reads as a self-loop
that the program does not contain.

The file's other RIL-b handlers already spell this (ULong)(Long)(Int)i2,
so this makes BRCTH agree with them.

+++ priv/guest_s390_toIR.c
@@ -5364,7 +5364,7 @@ s390_irgen_BRCTH(UChar r1, UInt i2)
 {
    put_gpr_w0(r1, binop(Iop_Sub32, get_gpr_w0(r1), mkU32(1)));
    if_condition_goto(binop(Iop_CmpNE32, get_gpr_w0(r1), mkU32(0)),
-                     guest_IA_curr_instr + ((ULong)(Long)(Short)i2 << 1));
+                     guest_IA_curr_instr + ((ULong)(Long)(Int)i2 << 1));
 
    return "brcth";
 }
